Refactor: exception 관련 refactoring
