--- src/main/java/com/sprint/mission/discodeit/controller/FileProcessingException.java
@@ -1,11 +1,0 @@
-package com.sprint.mission.discodeit.controller;
-
-public class FileProcessingException extends RuntimeException {
-    public FileProcessingException(String message) {
-        super(message);
-    }
-
-    public FileProcessingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
