+++ src/main/java/com/sprint/mission/discodeit/exception/FileProcessingException.java
@@ -1,0 +1,12 @@
+package com.sprint.mission.discodeit.exception;
+
+public class FileProcessingException extends RuntimeException {
+
+  public FileProcessingException(String message) {
+    super(message);
+  }
+
+  public FileProcessingException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
